pyproject.toml
[project]
name = "storer"
description = "A simple FastAPI example application"
dynamic = ["version"]
requires-python = ">=3.14"
readme = "README.md"
license = { file = "LICENSE.md"}
dependencies = [
    "alembic>=1.18.4",
    "bcrypt>=5.0.0",
    "bitaware>=0.2.0",
    "fastapi[standard]>=0.136.0",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.12.1",
    "sqlmodel>=0.0.38",
]

[project.urls]
"Homepage" = "https://github.com/wnowicki/storer"
"Source" = "https://github.com/wnowicki/storer"

[tool.setuptools]
py-modules = []

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=app --cov-report xml:coverage.xml --cov-report term"
testpaths = [
    "tests",
]
pythonpath = "."

[project.scripts]
my_app = "app.main:main"

[tool.coverage.run]

omit = [
    "*/__init__.py",
    "alembic/*",
]

[tool.ruff]

line-length = 100
extend-exclude = ["*.ipynb"]

[tool.ruff.lint]
# 1. Enable flake8-bugbear (`B`) rules, in addition to the defaults.
select = [
    "E", # All errors
    "F", # PyFlakes
    "B", # Bugbear
    "Q", # Quoting
    "I", # Isort
    "S", # flake8-bandit
    "N", # pep8-naming
    "A", # flake8-builtins
    "FAST", # FastAPI specific rules
    "C4", # flake8-comprehensions
    # "FIX", # Fixme
    "LOG", # flake8-logging
    "RET", # flake8-return
    "SIM", # flake8-simplify
    "C90", # mcCabe complexity
]

# 2. Avoid enforcing line-length violations (`E501`)
ignore = ["E501"]

# 3. Avoid trying to fix flake8-bugbear (`B`) violations.
unfixable = ["B"]

# 4. Ignore `E402` (import violations) in all `__init__.py` files, and in selected subdirectories.
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F403"]
"**/{tests,docs,tools}/*" = ["E402", "S101"]

[dependency-groups]
dev = [
    "pre-commit",
    "pytest",
    "pytest-cov",
    "ruff",
]
